# umit-mobilya-server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Registers a user and signs them in straight away.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Signs an existing user in.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The user behind the supplied Bearer token. Never includes the password.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/categories": {
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Every category, unfiltered. Public — the storefront reads it.",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creates a category.",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/filter": {
      "get": {
        "operationId": "CategoryController_filter",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Filters categories by name.\n\nReads the criteria from the request BODY of a GET, which is unusual and\ndeliberate: that is what the Express version did and what the frontend\nsends. Changing it to a query parameter is a breaking change for callers.",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/{id}": {
      "put": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Renames a category.",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletes a category. Answers with a message, not the deleted document.",
        "tags": [
          "categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ümit Mobilya API",
    "description": "Frontend istemcisi bu şemadan üretilir (`yarn gcl`). Bir uç burada görünmüyorsa istemcide de yoktur.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Login address. Stored lowercased and unique across users.",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Plain password, at least six characters. Never stored as given.",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "The signed-in user's id, as a bare hex string — not an object."
          },
          "token": {
            "type": "string",
            "description": "Signed JWT, valid for three days. Also set as an httpOnly `jwt` cookie on\nthe same response; the frontend stores this copy in localStorage."
          }
        },
        "required": [
          "user",
          "token"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Registered login address.",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Plain password, compared against the stored bcrypt hash."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable confirmation, safe to show to the user as-is."
          }
        },
        "required": [
          "message"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "description": "Login address, always lowercased. The password hash is never included."
          }
        },
        "required": [
          "_id",
          "email"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Always \"Token is valid\". Kept because the frontend store reads the envelope."
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Mongo id, as a hex string."
          },
          "name": {
            "type": "string",
            "description": "Display name shown in the category list."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO-8601 timestamp of when the category was added."
          }
        },
        "required": [
          "_id",
          "name",
          "createdAt"
        ]
      },
      "FilterCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Case-insensitive substring to match against the category name. Omit it to\nget every category back.",
            "maxLength": 120
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name shown in the category list, e.g. \"Koltuk Takımı\".",
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New display name for the category.",
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}